+++ spotipy/oauth2.py
@@ -21,7 +21,7 @@ import requests
 
 from spotipy.cache_handler import CacheFileHandler, CacheHandler
 from spotipy.exceptions import SpotifyOauthError, SpotifyStateError
-from spotipy.util import CLIENT_CREDS_ENV_VARS, get_host_port, normalize_scope
+from spotipy.util import CLIENT_CREDS_ENV_VARS, REQUESTS_SESSION, get_host_port, normalize_scope
 
 logger = logging.getLogger(__name__)
 
@@ -122,7 +122,7 @@ class SpotifyAuthBase:
 
     def __del__(self):
         """Make sure the connection (pool) gets closed"""
-        if isinstance(self._session, requests.Session):
+        if getattr(self, "_session", None) and isinstance(self._session, REQUESTS_SESSION):
             self._session.close()